+++ processor/src/test/java/dz/jtsgen/processor/renderer/module/ModuleGeneratorTest.java
@@ -1,0 +1,57 @@
+/*
+ * Copyright (c) 2017 Dragan Zuvic
+ *
+ * This file is part of jtsgen.
+ *
+ * jtsgen is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * jtsgen is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with jtsgen.  If not, see http://www.gnu.org/licenses/
+ *
+ */
+
+package dz.jtsgen.processor.renderer.module;
+
+import dz.jtsgen.processor.model.TypeScriptModel;
+import dz.jtsgen.processor.renderer.model.TypeScriptRenderModel;
+import org.junit.Test;
+
+import javax.annotation.processing.Filer;
+import javax.annotation.processing.Messager;
+import javax.annotation.processing.ProcessingEnvironment;
+
+import java.io.IOException;
+
+import static org.junit.Assert.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+public class ModuleGeneratorTest {
+
+    @Test
+    public void writeModule() throws Exception {
+        TypeScriptRenderModel model = new TypeScriptRenderModel(TypeScriptModel.newModelWithDefaultModule());
+        ProcessingEnvironment env = mock(ProcessingEnvironment.class);
+        Filer filer = mock(Filer.class);
+        Messager messager = mock(Messager.class);
+        when(env.getMessager()).thenReturn(messager);
+        when(filer.createSourceFile(any(),any(),any(),any())).thenThrow(new IOException("mocked IO Error"));
+        when(filer.createResource(any(),any(),any())).thenThrow(new IOException("mocked IO Error"));
+        when(env.getFiler()).thenReturn(filer);
+        ModuleGenerator testee = new ModuleGenerator(model,env);
+        testee.writeModule();
+
+        verify(messager).printMessage(any(),any());
+    }
+
+}